refactor(unit test): extract test stubs to repositories/stubs
