+++ back-end/src/modules/createCourse/CreateCourseService.spec.ts
@@ -1,26 +1,6 @@
-import { Classroom } from "../../entities/Classroom";
-import { Course } from "../../entities/Course";
-import { Teacher } from "../../entities/Teacher";
-import { ICoursesRepository } from "../../repositories/ICoursesRepository";
-import { ICreateCourseDTO } from "./CreateCourseDTO";
 import { CreateCourseService } from "./CreateCourseService";
+import { CoursesRepositoryStub } from "../../repositories/stub/CoursesRepositoryStub";
 
-class CoursesRepositoryStub implements ICoursesRepository {
-    teachers: Array<Teacher>;
-    classrooms: Array<Classroom>;
-
-    constructor() {
-        this.teachers = new Array<Teacher>();
-        this.classrooms = new Array<Classroom>();
-
-        this.teachers.push(new Teacher({ name: 'test name'}));
-        this.classrooms.push(new Classroom({ number: '101'}));
-    }
-
-    async save (data: ICreateCourseDTO): Promise<Course> {
-        return new Course({...data, teachers: this.teachers, classrooms: this.classrooms});
-    }
-}
 
 describe("Create course", () => {
     let sut: CreateCourseService;
